Move project data class to project file
