+++ novelwriter/gui/noveltree.py
@@ -109,7 +109,7 @@ class GuiNovelView(QWidget):
     def refreshTree(self):
         """Refresh the current tree.
         """
-        self.novelTree.refreshTree(rootHandle=self.theProject.data.getLastHandle("noveltree"))
+        self.novelTree.refreshTree(rootHandle=self.theProject.data.getLastHandle("novelTree"))
         return
 
     def clearProject(self):
@@ -122,7 +122,7 @@ class GuiNovelView(QWidget):
     def openProjectTasks(self):
         """Run open project tasks.
         """
-        lastNovel = self.theProject.data.getLastHandle("noveltree")
+        lastNovel = self.theProject.data.getLastHandle("novelTree")
         if lastNovel not in self.theProject.tree:
             lastNovel = self.theProject.tree.findRoot(nwItemClass.NOVEL)
 
@@ -319,7 +319,7 @@ class GuiNovelToolBar(QWidget):
     def _refreshNovelTree(self):
         """Rebuild the current tree.
         """
-        rootHandle = self.theProject.data.getLastHandle("noveltree")
+        rootHandle = self.theProject.data.getLastHandle("novelTree")
         self.novelView.novelTree.refreshTree(rootHandle=rootHandle, overRide=True)
         return
 
@@ -485,7 +485,7 @@ class GuiNovelTree(QTreeWidget):
             titleKey = selItem[0].data(self.C_TITLE, self.D_KEY)
 
         self._populateTree(rootHandle)
-        self.theProject.data.setLastHandle(rootHandle, "noveltree")
+        self.theProject.data.setLastHandle(rootHandle, "novelTree")
 
         if titleKey is not None and titleKey in self._treeMap:
             self._treeMap[titleKey].setSelected(True)
@@ -523,7 +523,7 @@ class GuiNovelTree(QTreeWidget):
             self._lastCol = colType
             self.setColumnHidden(self.C_EXTRA, colType == NovelTreeColumn.HIDDEN)
             if doRefresh:
-                lastNovel = self.theProject.data.getLastHandle("noveltree")
+                lastNovel = self.theProject.data.getLastHandle("novelTree")
                 self.refreshTree(rootHandle=lastNovel, overRide=True)
         return
